requirements.txt
absl-py==1.0.0
albumentations==1.1.0
astunparse==1.6.3
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
colorama==0.4.4
cycler==0.11.0
Cython==0.29.26
easydict==1.9
filelock==3.4.2
flatbuffers==2.0
fonttools==4.28.5
gast==0.4.0
gdown==3.10.1
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.6.0
idna==3.3
imageio==2.13.5
importlib-metadata==4.10.0
imutils
#insightface==0.5
joblib==1.1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
Markdown==3.3.6
matplotlib==3.5.1
networkx==2.6.3
numpy==1.21.5
oauthlib==3.1.1
onnx==1.10.2
onnxruntime==1.10.0
opencv-python==4.5.5.62
opencv-python-headless==4.5.5.62
opt-einsum==3.3.0
packaging==21.3
Pillow==9.0.0
prettytable==3.0.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
PySocks==1.7.1
python-dateutil==2.8.2
PyWavelets==1.2.0
PyYAML==6.0
qudida==0.0.4
requests==2.27.1
requests-oauthlib==1.3.0
retina-face==0.0.7
rsa==4.8
scikit-image==0.19.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
threadpoolctl==3.0.0
tifffile==2021.11.2
tqdm==4.62.3
typing_extensions==4.0.1
urllib3==1.26.8
wcwidth==0.2.5
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.7.0
